`CEventHandler` (#615)
* ctor and dtor

* fix `FastLoader::StartGame`

* `HandleEvents`

* Fix code

* `HandleEvents`

* `RegisterKill`

* `SetEventResponseTask`

* `ComputeAreaCodesResponse`

* `CTaskComplexUseEffect`

* Fix release build

* `ComputeAttractorResponse`

* fix COS_45`

* stub for `CPedGeometryAnalyser::CanPedJumpObstacle`

* Add CEntity* ctor to `CTaskComplexWalkRoundBuildingAttempt`

* `ComputeBuildingCollisionResponse`

* `ComputeCarUpsideDownResponse`

* fix `ComputeBuildingCollisionResponse`

* fix `ComputeCarUpsideDownResponse`

* fix  `ComputeChatPartnerResponse`

* `ComputeChatPartnerResponse`

* `ComputeCopCarBeingStolenResponse`

* `CTaskComplexPartnerGreet`

* `CAnimManager::StreamAnimBlock`

* `ComputeCreatePartnerTaskResponse`

* `ComputeDangerResponse`

* `ComputeDeadPedResponse`

* Fix `ComputeDeadPedResponse`

* `ComputeDeathResponse`

* `ComputeDontJoinGroupResponse`

* Fix `tHandlingData::m_nAnimGroup`

* Refactor `CTaskComplexSequence`

* `CTaskComplexKillCriminal`

* `CTaskSimpleAchieveHeading`: Add new ctor

* `CTaskComplexEnterCarAsPassenger`: add ctor default args

* `CTaskComplexDragPedFromCar`: add default args

* `CTaskComplexSequence`: Add `AddTasks`

* `GetCurrentEventType`

* `GetCurrentEventType`

* `0x4BCC7E`

* `ComputeFireNearbyResponse`

* `CTaskComplexSmartFleeEntity`: Rename `fleeEntity` to `fleeFrom`

* `ComputeGotKnockedOverByCarResponse`

* Fix `ComputeFireNearbyResponse`

* `CTaskComplexUseClosestFreeScriptedAttractor`

* `CTaskComplexUseClosestFreeScriptedAttractorRun` and `CTaskComplexUseClosestFreeScriptedAttractorSprint`

* `ComputeGunAimedAtResponse`

* format `CTaskComplexDie`

* `CWeaponInfo`: Refactor and add better docs

* `ComputeDamageResponse`

* `ComputeHighAngerAtPlayerResponse`

* `ComputeInWaterResponse`

* Add `CTaskInteriorUseInfo`

* `ComputeInteriorUseInfoResponse`

* `ComputeLowAngerAtPlayerResponse`

* `CTaskComplexEvasiveDiveAndGetUp`

* `CTaskComplexDiveFromAttachedEntityAndGetUp`

* eDIrection enum

* `ComputeLowHealthResponse`

* `ComputeObjectCollisionPassiveResponse`

* hook `ComputeLowHealthResponse`

* hook `ComputeLowAngerAtPlayerResponse`

* `ComputeObjectCollisionResponse`

* `ComputeOnEscalatorResponse`

* `ComputeOnFireResponse`

* `ComputePassObjectResponse`

* `ComputePedFriendResponse`

* `ComputePedSoundQuietResponse`

* `ComputePedThreatBadlyLitResponse`

* `CTaskComplexSeekEntityAiming`

* Add `eDrivebyStyle`

* `ComputePedThreatResponse`

* `ComputePedToChaseResponse`

* `ComputePedToFleeResponse`

* `ComputePersonalityResponseToDamage`

* `ComputePlayerCollisionWithPedResponse`

* `ComputePlayerWantedLevelResponse`

* `ComputePotentialPedCollideResponse`

* `ComputePotentialWalkIntoFireResponse`

* `ComputeReallyLowHealthResponse`

* `ComputeReviveResponse`

* `ComputeScriptCommandResponse`

* fix `CTaskComplexUseClosestFreeScriptedAttractorSprint`

* fix `CTaskComplexPartnerGreet`

* fix `CTaskComplexUseClosestFreeScriptedAttractorRun`

* Fix `CPed::GetWeaponSkill`

* hook `ComputeEventResponseTask`

* `ComputeSeenCopResponse`

* `ComputeSeenPanickedPedResponse`

* `ComputeSexyPedResponse`

* `ComputeSexyVehicleResponse`

* `ComputeShotFiredResponse`

* `ComputeShotFiredWhizzedByResponse`

* `ComputeSignalAtPedResponse`

* `ComputeSpecialResponse`

* Fix code for CI

* `ComputeVehicleCollisionResponse`

* `ComputeVehicleDamageResponse`

* `ComputeVehicleDiedResponse`

* `ComputeVehicleOnFireResponse`

* `ComputeVehiclePotentialCollisionResponse`

* `ComputeVehiclePotentialPassiveCollisionResponse`

* `ComputeVehicleToStealResponse`

* `ComputeWaterCannonResponse`

* `ComputePedEnteredVehicleResponse`

* Fix formatting

* Addendum `ComputeVehicleCollisionResponse`

* Workaround for crash in `CAESound::UpdatePlayTime(`

* Fix `CEventHandler::ComputeOnFireResponse`

* Formatting

* Small refactor

* `CEventGunAimedAt`: Rename `m_ped` to `m_AimedBy`

* .clang-format

* Fix bug in `CWeapon::GenerateDamageEvent`

* Fix build

* Fix `CWeapon::GenerateDamageEvent`

* Fix `CTaskSimpleBeHit::ProcessPed`

* Move `CAEPedSpeechAudioEntity::InjectHooks` to where it should be

* Fix `CEventHandler::HandleEvent`

* Disable `DoPathSearch` hook

* Fix build after merge

* fix merge

* Add `CTask::CloneIfIs`

* Refactor `CPedIntelligence::FlushImmediately`

* Make member variables private + change names to `m_CamelCase`

* Simplify logic

* Make `CEventHandlerHistory` members private

* Fix build

+++ source/game_sa/Tasks/TaskTypes/TaskSimpleGunControl.cpp
@@ -21,6 +21,7 @@ void CTaskSimpleGunControl::InjectHooks() {
 // because that took pointers instead for the CVector's, but that's stupid, because
 // they just 0 inited it otherwise, thus we can just default init the args
 // instead of passing nullptr, and it's going to be the same effect.
+// so, tldr, instead of `nullptr` for the CVector&'s pass in `{}`
 CTaskSimpleGunControl::CTaskSimpleGunControl(CEntity* targetEntity, CVector const& targetPos, CVector const& moveTarget, eGunCommand firingTask, int16 burstAmmoCnt, int32 leisureDurMs) :
     m_targetPos{targetPos},
     m_moveTarget{moveTarget},
@@ -109,7 +110,7 @@ bool CTaskSimpleGunControl::ProcessPed(CPed* ped) {
             const auto useGun = new CTaskSimpleUseGun{
                 m_targetEntity,
                 m_targetPos,
-                (uint8)eGunCommand::AIM,
+                eGunCommand::AIM,
                 (uint16)m_burstAmmoCnt,
                 m_aimImmidiately
             };
@@ -246,7 +247,7 @@ bool CTaskSimpleGunControl::ProcessPed(CPed* ped) {
 
         if (!m_leisureDurMs && useGunTask) {
             m_firingTask = END_LEISURE;
-            useGunTask->ControlGun(ped, m_targetEntity, (int8)END_LEISURE);
+            useGunTask->ControlGun(ped, m_targetEntity, END_LEISURE);
         }
     }
 
@@ -273,9 +274,8 @@ bool CTaskSimpleGunControl::ProcessPed(CPed* ped) {
     }();
 
     if (useGunTask) {
-        useGunTask->ControlGun(ped, m_targetEntity, (int8)nextGunCmd);
+        useGunTask->ControlGun(ped, m_targetEntity, nextGunCmd);
     }
-
 
     CVector pedToTarget{};
     if (!GetPedToTarget(ped, pedToTarget)) {
